docs(core): the pages still described files this stack deleted
A page pointing at a file that is gone is worse than no page: it reads as
current. Six of them did.

  writing-a-provider.md   internal/core/classify.go translating a provider
                          failure onto core.RetryableError and
                          core.ContextExceededError. All three are gone —
                          ai.IsRetryable and ai.IsContextExceeded answer
                          directly, which is the point the guide was making.
  llm.md                  the same file, tagging an error as it leaves the
                          stream. That is ai.StreamError now.
  core.md                 aiconv.go and classify.go as core's two converters,
                          and agent_impl.go as its single implementation file.
                          There is one converter left (ToAITools, in tool.go,
                          because San holds a schema without a Run) and two
                          implementation files.
  compaction.md           ThinkAct and compact in agent_impl.go. Compaction is
                          preStep, onInferError and shorten, in exchange.go.
  data-flow.md / .zh.md   a markdown link to internal/core/agent_impl.go — a
                          link that resolved to nothing.

And the test file kept the name of the source file it was named after.
agent_impl.go became run.go and exchange.go where the loop moved;
agent_impl_test.go did not follow, so twelve tests sat in a file named for
something that no longer exists. Split the way the source is: run_test.go is
the mailbox — the signals, the drains, the interrupt latch — and
exchange_test.go is one exchange. TestStreamInferReportsAnUnreachableModel was the last
thing still calling the placeholder client unreachable.
